Almost Finished
Added parameter queuing button, a sample queueing option and the skeleton of a sample removal option. Need to just integrate in the processThread function and let it loop across the set of samples and their configurations. The outputs need to be organized into folders for each sample with an accompanying json for the parameters of each variation.