Integrated RACC properly
A function was added to convert 8-bit images to RGB for RACC due to the required dimensionality. This could be reviewed in future as this dimensionality reduction might be redundant but could feed into further dependencies within the function. This recent function is a band-aid for the problem but it is likely not optimal. RACC returns results that are close to the Union of the two channels tested with and the colocalized voxels align with union overlaps. More testing with this pipeline should be performed