Update AutoHysteresis.py
Further improved the high threshold finding method. Currently it back tracks to acquire voxels that were missed due to similar intensity bridges away from the core structures. Currently the average of the change in voxels between intensity steps is used to select the estimated intensity but this is being tested