Testing more efficient loops for high threshold and sample_checker for result validation
Added a variant of the nested for loops to rather use nditer in the high threshold calculation method in the hopes that it is more efficient. Sample_checker compares automated threshold with the less effective Otsu to determine if sufficient voxels have been retained. If retained voxels are less than a fraction of Otsu then disqualified. Also a method to add ratios of generated Gaussian noise but value is to be determined