Update to noise addition and thresholding with data visualisation
sample_checker was updated to correctly add noise to deconvolved samples which are then preprocessed. With this correction regarding noise a new variant of the testing() function does not add new noise and instead thresholds the noisy samples. The results from the thresholding are now pushed to a JSON file. ResultExplorer has been started and will use panda dataframes to explore the results of the thresholding