Added dictionary and list handling for data plotting
ResultExplorer can now dynamically determine if a Dataframe row contains a list and if so it can separate the lists into new columns of a similar name. Dictionary handling is also improved and getting the mean of the data is now optional. BetaThresholdingClean is generating new data but adding metrics for the percentage of voxels per intensity above low and the change in voxel count between these intensities.