Updated Thresholding metric organization and added graphing functions
For the latest set of graphs the json recording of BetaThresholdingClean was adjusted. These results are compiled in Graphmaker which is to create a different set of graphs than ResultExplorer and won't get stuck with Class related issues as these functions and not made with the intention of  definite future use.