BetaThresholdingCleanDupe added
BetaThresholdingCleanDupe added. Originally a duplicate of BetaThresholdingClean which point to a different directory for input and output so that the same script could be run simultaneously to reduce total time to generate data. Experimental but added a while loop outside of the intensity range to address potential oscillating intensity bridges and changed fixed intensity steps so as not to cut-out voxels between lowest intensity range value and actual low threshold value. Experimental