New functionality for data generation and graphing
An intersection function was added to the graphing file to determine the intensity between two points. This was made to improve the threshold selection. The iterate_through_hysteresis function in BetaThresholdingClean and Dupe will all the preview of the threshold results of a varying high threshold and the total voxels for each high threshold. The looping condition and starting density have been skipped along with the 'fixed' configuration as they provide unusable or negligible results. The precision range has been increased to depict a trend in high threshold values depending on the resolution step size.