Updated GraphMaker and both BetaThresholdingClean
For the BetaThresholdingClean the testing_knee function was fixed as the first know returned was not the actual first knee and now can be used if required. The generation of a Hysteresis preview was fixed. GraphMaker had an image stitching method added to stitch thresholded images together for review. A histogram generation method was added which generates a histogram based on json data. intensity_graphs3 was changed to evaluate a list of blocked precision values for comparison to prior iterations.