Improvements for cumulative iterative hysteresis method
Added a low threshold selection method which uses kneedle in combination with the normal hist counts and the log of the hist counts for two separate low thresholds. If the Normal count low threshold is less than the Otsu threshold then the log is used. Acquires a better low threshold value. Created a method to rescale the gradient graphs such that slopes closer to zero are weighted highly compared to higher gradient scalar values. Using this a mean is calculated over a descending window to determine a left biased high intensity threshold.