Update GraphMaker.py
Added a biased mean test to verify behaviour of gaussian rescaling. Adjusted Gaussian rescaling in terms of approximating a left biased centroid of the slope histogram to determine the high threshold. Added the tunable method which attempts to determine the high threshold by taking the first log slope intersection at a specific value to the left of the first local maxima. There can be multiple local maxima and thus find_cluster and refine_cluster are designed to determine the "true" local maxima to use for the high threshold selection.