Voxel weighting, lines and timeframe splitting
Voxel weighting in the rescale_gaussian has improved clarity and the power of the voxel weights can be adjusted during the call. Gradient lines and their intersections where calculated for the histograms in an investigation for potential metrics in histogram crest selection (experimental). Time splitter is a script made to separate time dimension samples and to removed pointless rgb channels. The HysteresisThresholding.ipynb now stores the current low and high threshold at every update of the sliders.