Added new files and setup batch processing
BetaThresholdingCleanDupe was adjusted so that it will run low theshold and high threshold tests on all listed data collections and record the results. Gaussian Blurring was added to the Synthetic Data generation for an improved gradient across the synthetic organelle volume.
CompleteDataReader visualises the cumulative Hysteresis histogram and annotates the centroid and Yen threshold intensities for evaluation. csvToPandasExplore summarises the variation caused be the individual Voxel Scaling power and the logistic function Steepness parameters on the determined high threshold.