Improved Volume generation efficiency and started building validator
Improved the runtime of the volume generation function (now takes half the time). A function to enable the generation of multiple complex structures in the image is in development. A ThresholdingValidation python notebook is in development which will be used for validation by Sholto and Ben to validate the results of the automatic system