Corrected Log for low thresholding and fixed rgb overlay generation
Changed the log of the gradient distribution from a statement where "y=log(x) if x is not 0 else 0" which meant y(0) = y(1) = 0. It was changed to be y(x) = log(x+1) since x >>> 1. The red over gray overlay for the mip was fixed for the thresholded image as the grey was the thresholded intensity instead of the original intensity in grey.