Added RatingEvaluation and VarianceSeparation
RatingEvaluation is a script that will attempt to quantify the differences in the thresholded images (experts to auto) in terms of numeric thresholds and thresholded images & the expert ratings. Currently this just extracts the thresholds, weightings and sources the images.

VarianceSeparator is a side project for curiousity. It currently aims to visualise the variance across independent structures with n-by-n neighbourhoods that stride by 1. View_as_windows is currently the best for segmenting. Further code needs to be made to communicate changes back to structure from windows and to minimise resource consumption.