Update VarianceSeparator.py
Added variance mapping with square windows (currently 3x3) with centre pixel checking (centre pixel must be non-zero) and background windows (all zero pixels) are also ignored. Currently the local variance for each pixel is measured but they are discrete and a measure that takes a continuous space perspective of the variance is needed to find drops in intensity between relatively higher values.