Updated SystemAnalysis and created a csv to json scraper
ratingsCSVscraper was created to extract the rating information from a .csv file and save it as a .json file (for Rensu's ratings). SystemAnalysis.py has added functionality to load in the expert thresholds and ratings for evaluation in future. The compilation of the expert thresholds was established and a method to compare between the automatic low threshold and the expert ones is being developed (_compare_with_experts)