Update SystemAnalysis.py
Completed the extraction and calculation of the low thresholds with respect to the expert results. These threshold results still need a sufficient quantification to describe their applicability (currently closeness is sufficient for the low threshold). The impact relative to the image difference is being developed in _image_analysis and _structure_overlap