Update SystemAnalysis.py
Created _image_change_calculation which will be used to calculate the changes between a subject image and target image as the threshold parameters of the subject image approach that of the target image in iteration.

Structure overlapping was changed so that instead of adding the number of independent overlaps, between a pair of structures that each correspond to the subject and target, the shared overlap volume is stored instead. The number of independent overlaps between two specific structures is of marginal interest. The mapping to the array indexes will be used to match back to the relevant structures. Structures in each image that map to nothing are recorded as well. Must change to only record the fully excluded subject image structures