Update SystemAnalysis.py
Added the subject image detail extraction which will be used to determine the difference of the subject structures from the reference image structures. The nested function _overlap_ratios attempts to calculate the overlapped volume versus the complete volume for each structure vector-wise. The matching of reference structures to subject structures is still being developed.