Update SystemAnalysis.py
Added in a bridge structure example. _consolidate_prior_structures was created which attempt to map new image structures (post transformation/thresholding) to previous image transformations/thresholds. This is used to evaluate the change in an image as it's parameters tend towards the expert thresholds and deal with the structure join/split issue