Update SystemAnalysis.py
Adjusted overlap ratio to take in the ratio of overlap/im1 * overlap/im2 to avoid cases where the ratio is 1 but the volumes do not match (occurs when a smaller volume is entirely contained within another). Added distance_from_target which is a function that measures the overlap ratio between a subject image and a target image for all mutations of the subject image where it tends towards the target image. To balance out excluded structures in the subject image (not shared by the target image) a secondary mapping of the overlap ratio between the mutating subject image and the original subject image is determined.