Tested vectorized IHH
Tested vectorized IHH. With a max_mem of less than 1 the efficiency improvement are minimal of not detrimental. Currently the prior parameter demands a lot of memory and penalizes the memory available thus harming efficiency. The number of operations required for this can be demanding and even though they reduce the for loop steps the iteration count needs to meet a certain threshold for improvements