Added a potentially more efficient centroid calculation, independent structure counting and functions to visualize these
In CleanThresholder:
A potentially more efficient weighted centroid
calculation function was developed but the execution time has yet to be measured for efficiency.

The independent structure counting was added to the IHH generation in hopes of providing more data to determine effective voxel weighting (left bias) with some measure of nuance.

The weighted centroid calculation was corrected as the IHH received by the function was reversed for some reason so it is reversed within the function currently.

In SystemAnalysis:
The visualization of the independent structure counting is performed.