For voxel structure analysis
Removed reverse in centroid calc for CleanThresholder as there is erroneous reversing somewhere and it was interfering with the algorithm. In SystemAnalysis a number of transformations of the voxel structure sizes is explored and visualized and the effect of these on the voxel weights is calculated and printed for analysis.