Added new functions and arguments for structure infor
Added new functions to SystemAnalysis and built on existing ones to further investigate the potential uses of the structure size (and count) related information to improve the high threshold calculation.

The window centroid function in CleanThresholder was adjusted such that an optional argument can be used to apply an extra weighting distribution to the voxel weight without affecting the window mass percentage scaling.