Added intensity centroid and other metrics
Added in the calculation of the intensity centroid without windowing. The determination of the IHH using an expert low threshold and an auto low threshold was the most recent addition.

Tests regarding the reverse windowing (fixed at max intensity) and the representation of automatic and expert thresholds by rgb images was performed